Vehicle override lights & vehicle fix

+++ SlipeServer.Server/PacketHandling/Builders/AddEntityPacketBuilder.cs
@@ -152,7 +152,7 @@ public class AddEntityPacketBuilder
             null, element.AreCollisionsEnabled, element.IsCallPropagationEnabled, new CustomData(), element.Name, element.TimeContext,
             element.Position, element.Rotation, element.Model, element.Health, (byte)element.BlownState, element.Colors.AsArray(), element.PaintJob, element.Damage, element.Variants.Variant1,
             element.Variants.Variant2, element.TurretRotation, element.AdjustableProperty, VehicleConstants.DoorsPerVehicle[(VehicleModel)element.Model] > 0 ? element.DoorRatios.ToArray() : Array.Empty<float>(), MapVehicleUpgrades(element.Model, element.Upgrades), element.PlateText, 
-            element.OverrideLights, element.IsLandingGearDown, element.IsSirenActive, element.IsFuelTankExplodable, element.IsEngineOn, element.IsLocked, 
+            (byte)element.OverrideLights, element.IsLandingGearDown, element.IsSirenActive, element.IsFuelTankExplodable, element.IsEngineOn, element.IsLocked, 
             element.AreDoorsDamageProof, element.IsDamageProof, element.IsFrozen, element.IsDerailed, element.IsDerailable, element.TrainDirection == Elements.Enums.TrainDirection.Clockwise, element.IsTaxiLightOn, 
             element.Alpha, element.HeadlightColor, element.Handling, element.Sirens
         );
